iommu: iommu_do_domctl change initial ret value.
There are some code wrapped in the blocks #ifdef..#endif, inside of the blocks
the ret value is compared with '-ENOSYS' and if it is, start the code execution.
It means that in the code below, function iommu_do_dt_domctl can not be executed
if CONFIG_HAS_PCI not defined

int ret = -ENOSYS;

if ( !is_iommu_enabled(d) )
    return -EOPNOTSUPP;

ret = iommu_do_pci_domctl(domctl, d, u_domctl);

if ( ret == -ENOSYS )
   ret = iommu_do_dt_domctl(domctl, d, u_domctl);

So, it is reasonable to have '-ENOSYS' as initial value.

Signed-off-by: Ihor Usyk <ihor_usyk@epam.com>